pyproject.toml
[tool.poetry]
name = "lglutide"
version = "0.1.0"
description = ""
authors = ["kamal-shrestha <kamalandshrestha@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
notebook = "^6.5.3"
torch = "^1.13.1"
torchvision = "^0.14.1"
pandas = "^1.5.3"
pre-commit = "^3.1.1"
gdown = "^4.6.4"
tensorboard = "^2.12.0"
scikit-learn = "^1.2.2"
ipykernel = "^6.21.3"
Flask = "^2.2.3"
Flask-Cors = "^3.0.10"
opencv-python = "^4.7.0"
watermark = "^2.3.1"
python-dotenv = "^1.0.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
